pyproject.toml
[project]
name = "pcb-maker"
version = "0.1.0"
description = "Naive Gerber + drill to Snapmaker 2.0 G-code prototype"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygerber>=2.4.3",
    "pytest==8.2.2",
    "pyyaml==6.0.2",
    "shapely==2.0.4",
]

[project.scripts]
pcb-maker = "main:cli"
